requirements.txt
langchain
langchain-core
langchain-huggingface 
langgraph
streamlit   
python-dotenv